requirements.txt
# Core dependencies
python==3.10

numpy
pandas
opencv  # Mac
#opencv-python  # windows
scipy>=1.7.0,<1.14.0

# Tracking
filterpy>=1.4.5
motmetrics>=1.2.0

# Deep learning models
ultralytics # YOLO
pytorch
torchvision
open-clip-torch # CLIP

# Optional (for other baselines)
transformers>=4.30.0,<5.0.0  # For CLIP, Grounding DINO
# groundingdino  # Install separately if needed
# fairseq  # For DINO features

# Utilities
tqdm>=4.65.0
matplotlib>=3.5.0,<4.0.0
Pillow>=9.0.0,<10.0.0
pyyaml>=6.0
Ï
